Remove Redundant File, Separate Networking into own namespace, Separate FileInfo
