+++ Boneworks/BWUtil.cs
@@ -10,6 +10,7 @@ using UnityEngine.Rendering;
 using static UnityEngine.Object;
 
 using MultiplayerMod.Structs;
+using MultiplayerMod.Networking;
 
 namespace MultiplayerMod
 {
