+++ MultiplayerMod.csproj
@@ -291,6 +291,7 @@
     <Compile Include="Discord\StoreManager.cs" />
     <Compile Include="Boneworks\EnemyPoolManager.cs" />
     <Compile Include="Source\Extras\SpecialUsers.cs" />
+    <Compile Include="Source\FileInfo.cs" />
     <Compile Include="Source\Main.cs" />
     <Compile Include="Source\Core\MultiplayerUI.cs" />
     <Compile Include="Source\Networking\NetworkMessages.cs" />
