+++ Source/Core/Client.cs
@@ -19,6 +19,7 @@ using StressLevelZero.Pool;
 using StressLevelZero.AI;
 
 using MultiplayerMod.Structs;
+using MultiplayerMod.Networking;
 using MultiplayerMod.Representations;
 
 namespace MultiplayerMod.Core
