+++ Source/Core/Server.cs
@@ -18,6 +18,7 @@ using UnityEngine.SceneManagement;
 using static UnityEngine.Object;
 
 using MultiplayerMod.Structs;
+using MultiplayerMod.Networking;
 using MultiplayerMod.Representations;
 
 namespace MultiplayerMod.Core
