+++ Source/FileInfo.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace MultiplayerMod
+{
+    public static class FileInfo
+    {
+        public const string Name = "Multiplayer Mod";
+        public const string Author = "Someone Somewhere";
+        public const string Company = "Lava Gang";
+        public const string Version = "0.11.0";
+        public const string DownloadLink = "https://discord.gg/2Wn3N2P";
+    }
+
+}
